before the artist
