+++ btc_cycles/artists/__init__.py
@@ -1,1 +1,5 @@
 """artist submodule"""
+
+from .artist import Artist
+
+__all__ = ["Artist"]
